+++ generate_matrix.hpp
@@ -47,5 +47,5 @@
 #endif
 #include "HPC_Sparse_Matrix.hpp"
 
-void generate_matrix(int nx, int ny, int nz, HPC_Sparse_Matrix **A, double **x, double **b, double **xexact);
+void generate_matrix(int nx, int ny, int nz, HPC_Sparse_Matrix **A, float **x, float **b, float **xexact);
 #endif
